pyproject.toml
[project]
name = "memora"
version = "0.1.0"
description = "Terminal-based personal memory assistant"
requires-python = ">=3.10"
dependencies = [
    "textual",
    "fastembed",
    "rich",
    "typer",
    "langchain",
    "langchain-core",
    "langchain-openai",
    "langchain-text-splitters",
    "sentence-transformers",
    "faiss-cpu",
    "rank-bm25",
    "pypdf",
    "python-docx",
    "pydantic",
    "pydantic-settings",
    "python-dotenv",
    "numpy",
]

[project.scripts]
memora = "memora.cli:app"

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["src/memora"]
